Pass variant into <BlockToolbar />
The variant is used to style dropdown borders. All block toolbars are assumed to have a variant of "toolbar" which adds a dark black border to dropdowns.

I think it's worth revisiting this idea, as the block toolbar popover styles are not assumed for all toolbars, but this is the system that's currently in place, so I'm rolling with it for now.

If we could pass the variant in via the Popover instead, that would be ideal, and then remove the assumed toolbar context from the dropdowns.