Remove small white underline from parent block button
The background for the button was applied to the parent wrapper, and the wrapper also had a border without any border color, so it was bleeding past the button. Removing that file fixes it and doesn't seem to cause any issues. The applied style are provided by the button component now.